fish: add base directories for `cd` command
While trying to adopt @mrnugget's [`cdr`][1] shell alias which I found
very useful I came up with an approach that works even better for me.

Extending `CDPATH` environment variable with folders where I keep all
the projects I work on allows to jump not only to current project's
root, but to any (and not just root).

In combination with shell's autosuggestion abilities it enables
jumping around with almost the same amount of keystrokes as `cdr`, but
doesn't restrict it to a context of the current repository.

[1]: https://twitter.com/thorstenball/status/1223218245592911878